requirements.txt
flask
flask-login
flask-sqlalchemy
openai
reportlab
werkzeug
pdfplumber
python-docx